Add Lexeme type alias and simplify Lexeme type
Add Lexeme type alias and simplify Lexeme type. The main benefit of
adding a type alias for Lexeme is that I was able to change the
underlying type in 1 spot instead of many. This is super useful when I'm
not sure what the type should be and it allows me to change it easily.