Remove Position type and just store the offset
Some token implementations store the location as two numbers: the offset
from the beginning of the source file to the beginning of the lexeme, and
the length of the lexeme. The scanner needs to know these anyway, so
there’s no overhead to calculate them.

An offset can be converted to line and column positions later by looking
back at the source file and counting the preceding newlines. That sounds
slow, and it is. However you only need to do it when you need to actually
display a line and column to the user. Most tokens never appear in an error
message. For those, the less time you spend calculating position information
ahead of time, the better.

Minor:
- Remove Lexeme type.
- Add scanner.curr_lexeme helper.
- Catch whitespace in the default match block of scanner.scan_tokens.